moved more code to the cpp
