+++ src/snopt7.cpp
@@ -78,6 +78,140 @@ extern "C" {
 
 namespace pagmo
 {
+namespace detail
+{
+// We use this to ensure deleteSNOPT is called also if exceptions occur.
+template <typename snProblem>
+struct sn_problem_raii {
+    sn_problem_raii(snProblem *p, char *a, char *b, int n,
+                    std::function<void(snProblem *, char *, char *, int)> &snInit,
+                    std::function<void(snProblem *)> &deleteSNOPT)
+        : m_prob(p), m_deleteSNOPT(deleteSNOPT)
+    {
+        snInit(p, a, b, n);
+    }
+    ~sn_problem_raii()
+    {
+        m_deleteSNOPT(m_prob);
+    }
+    snProblem *m_prob;
+    std::function<void(snProblem *)> &m_deleteSNOPT;
+};
+
+inline void snopt_fitness_wrapper(int *Status, int *n, double x[], int *needF, int *nF, double F[], int *needG,
+                                  int *neG, double G[], char cu[], int *lencu, int iu[], int *leniu, double ru[],
+                                  int *lenru)
+{
+    (void)n;
+    (void)cu;
+    (void)lencu;
+    (void)ru;
+    (void)lenru;
+    (void)leniu;
+    // First we recover the info we have hidden in the workspace
+    auto &info = *(static_cast<detail::user_data *>(static_cast<void *>(iu)));
+    auto &verb = info.m_verbosity;
+    auto &log = info.m_log;
+    auto &f_count = info.m_objfun_counter;
+    auto &p = info.m_prob;
+    auto &dv = info.m_dv;
+    // We copy the decision vector into the vector_double
+    std::copy(x, x + p.get_nx(), dv.begin());
+    // We try to call the UDP fitness and gradient
+    try {
+        if (*needF > 0) {
+            auto fit = p.fitness(dv);
+            for (size_t i = 0u; i < static_cast<size_t>(*nF); ++i) {
+                F[i] = fit[i];
+            }
+
+            if (verb && !(f_count % verb)) {
+                // Constraints bits.
+                const auto ctol = p.get_c_tol();
+                const auto c1eq
+                    = detail::test_eq_constraints(fit.data() + 1, fit.data() + 1 + p.get_nec(), ctol.data());
+                const auto c1ineq = detail::test_ineq_constraints(fit.data() + 1 + p.get_nec(), fit.data() + fit.size(),
+                                                                  ctol.data() + p.get_nec());
+                // This will be the total number of violated constraints.
+                const auto nv = p.get_nc() - c1eq.first - c1ineq.first;
+                // This will be the norm of the violation.
+                const auto l = c1eq.second + c1ineq.second;
+                // Test feasibility.
+                const auto feas = p.feasibility_f(fit);
+
+                if (!(f_count / verb % 50u)) {
+                    // Every 50 lines print the column names.
+                    print("\n", std::setw(10), "objevals:", std::setw(15), "objval:", std::setw(15),
+                          "violated:", std::setw(15), "viol. norm:", '\n');
+                }
+                // Print to screen the log line.
+                print(std::setw(10), f_count + 1u, std::setw(15), fit[0], std::setw(15), nv, std::setw(15), l,
+                      feas ? "" : " i", '\n');
+                // Record the log.
+                log.emplace_back(f_count + 1u, fit[0], nv, l, feas);
+            }
+
+            // Update the counter.
+            ++f_count;
+        }
+
+        if (*needG > 0 && p.has_gradient()) {
+            auto grad = p.gradient(dv);
+            for (size_t i = 0u; i < static_cast<size_t>(*neG); ++i) {
+                G[i] = grad[i];
+            }
+        }
+    } catch (...) {
+        *Status = -100; // signals to snopt7 that things went south and it should stop.
+        info.m_eptr = std::current_exception();
+    }
+}
+
+} // namespace detail
+
+// Init of the statics data
+snopt7::mutex_t snopt7::m_library_load_mutex;
+
+const snopt7::result_map_t snopt7::m_results
+    = {{0, "None"},
+       {1, "Finished successfully - optimality conditions satisfied"},
+       {2, "Finished successfully - feasible point found"},
+       {3, "Finished successfully - requested accuracy could not be achieved"},
+       {5, "Finished successfully - elastic objective minimized"},
+       {6, "Finished successfully - elastic infeasibilities minimized"},
+       {11, "The problem appears to be infeasible - infeasible linear constraints"},
+       {12, "The problem appears to be infeasible - infeasible linear equality constraints"},
+       {13, "The problem appears to be infeasible - nonlinear infeasibilities minimized"},
+       {14, "The problem appears to be infeasible - linear infeasibilities minimized"},
+       {15, "The problem appears to be infeasible - infeasible linear constraints in QP subproblem"},
+       {16, "The problem appears to be infeasible - infeasible nonelastic constraints"},
+       {21, "The problem appears to be unbounded - unbounded objective"},
+       {22, "The problem appears to be unbounded - constraint violation limit reached"},
+       {31, "Resource limit error - iteration limit reached"},
+       {32, "Resource limit error - major iteration limit reached"},
+       {33, "Resource limit error - the superbasics limit is too small"},
+       {34, "Resource limit error - time limit reached"},
+       {41, "Terminated after numerical difficulties - current point cannot be improved"},
+       {42, "Terminated after numerical difficulties - singular basis"},
+       {43, "Terminated after numerical difficulties - cannot satisfy the general constraints"},
+       {44, "Terminated after numerical difficulties - ill-conditioned null-space basis"},
+       {45, "Terminated after numerical difficulties - unable to compute acceptable LU factors"},
+       {51, "Error in the user-supplied functions - incorrect objective derivatives"},
+       {52, "Error in the user-supplied functions - incorrect constraint derivatives"},
+       {56, "Error in the user-supplied functions - irregular or badly scaled problem functions"},
+       {61, "Undefined user-supplied functions - undefined function at the first feasible point"},
+       {62, "Undefined user-supplied functions - undefined function at the initial point"},
+       {63, "Undefined user-supplied functions - unable to proceed into undefined region"},
+       {71, "User requested termination - terminated during function evaluation"},
+       {74, "User requested termination - terminated from monitor routine"},
+       {81, "Insufficient storage allocated - work arrays must have at least 500 elements"},
+       {82, "Insufficient storage allocated - not enough character storage"},
+       {83, "Insufficient storage allocated - not enough integer storage"},
+       {84, "Insufficient storage allocated - not enough real storage"},
+       {91, "Input arguments out of range - invalid input argument"},
+       {92, "Input arguments out of range - basis file dimensions do not match this problem"},
+       {141, "System error - wrong number of basic variables"},
+       {142, "System error - error in basis package"}};
 
 std::vector<char> snopt7::s_to_C(const std::string &in)
 {
@@ -86,12 +220,9 @@ std::vector<char> snopt7::s_to_C(const std::string &in)
     return retval;
 }
 
-snopt7::snopt7(bool screen_output, std::string snopt7_c_library,
-               unsigned minor_version)
+snopt7::snopt7(bool screen_output, std::string snopt7_c_library, unsigned minor_version)
     : m_snopt7_c_library(snopt7_c_library), m_minor_version(minor_version), m_integer_opts(), m_numeric_opts(),
-      m_screen_output(screen_output), m_verbosity(0), m_log()
-{
-}
+      m_screen_output(screen_output), m_verbosity(0), m_log(){};
 
 /// Evolve population.
 /**
@@ -248,7 +379,7 @@ std::string snopt7::get_extra_info() const
     } else {
         stream(ss, "\n\tScreen output: (snopt7)");
     }
-    stream(ss, "\n\tLast optimisation return code: ", detail::snopt_statics<>::results.at(m_last_opt_res));
+    stream(ss, "\n\tLast optimisation return code: ", m_results.at(m_last_opt_res));
     stream(ss, "\n\tIndividual selection ");
     if (boost::any_cast<population::size_type>(&m_select)) {
         stream(ss, "idx: ", std::to_string(boost::any_cast<population::size_type>(m_select)));
@@ -423,7 +554,7 @@ population snopt7::evolve_version(population &pop) const
     // We then try to load the library at run time and locate the symbols used.
     try {
         // Here we import at runtime the snopt7_c library and protect the whole try block with a mutex
-        std::lock_guard<std::mutex> lock(detail::snopt_statics<>::library_load_mutex);
+        std::lock_guard<std::mutex> lock(m_library_load_mutex);
         boost::filesystem::path path_to_lib(m_snopt7_c_library);
         if (!boost::filesystem::is_regular_file(path_to_lib)) {
             pagmo_throw(std::invalid_argument, "The snopt7_c library path was constructed to be: "
@@ -638,7 +769,7 @@ We report the exact text of the original exception thrown:
                             xmul.data(), F.data(), Fstate.data(), Fmul.data(), &nS, &nInf, &sInf);
 
     if (m_verbosity > 0u) {
-        print("\n", detail::snopt_statics<>::results.at(m_last_opt_res), "\n");
+        print("\n", m_results.at(m_last_opt_res), "\n");
     }
     // ------- We reinsert the solution if better -----------------------------------------------------------
     // Store the new individual into the population, but only if it is improved.
